Import only defined models or all
