+++ src/Console/Commands/FlushCommand.php
@@ -12,7 +12,7 @@ final class FlushCommand extends Command
     /**
      * {@inheritdoc}
      */
-    protected $signature = 'scout:flush {searchable? : The name of the searchable}';
+    protected $signature = 'scout:flush {searchable?* : The name of the searchable}';
 
     /**
      * {@inheritdoc}
@@ -22,11 +22,14 @@ final class FlushCommand extends Command
     /**
      * {@inheritdoc}
      */
-    public function handle(SearchableListFactory $factory): void
+    public function handle(): void
     {
         $command = $this;
-        $searchables = (array) $command->argument('searchable');
-        $factory->make()->each(function ($searchable) {
+        $searchableList = collect($command->argument('searchable'))->whenEmpty(function () {
+            $factory = new SearchableListFactory(app()->getNamespace(), app()->path());
+            return $factory->make();
+        });
+        $searchableList->each(function ($searchable){
             $searchable::removeAllFromSearch();
             $this->output->success('All ['.$searchable.'] records have been flushed.');
         });
